+++ packages/lib/src/propTransform/index.ts
@@ -1,0 +1,1 @@
+export { propTransform, PropTransform, PropTransformDecorator } from "./propTransform"